requirements.txt
comet_ml==3.45.0
easydict==1.13
einops==0.8.1
loguru==0.7.2
matplotlib==3.10.3
networkx==3.3
numpy==2.3.1
open3d==0.18.0
Pillow==11.3.0
pyfqmr==0.2.0
pytorch3d.egg==info
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.16.0
setuptools==69.5.1
Shapely==2.1.1
sortedcollections==2.1.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
termcolor==3.1.0
timm==1.0.16
torch==2.3.1
torch==2.3.0
torchvision==0.18.1
tqdm==4.66.4
triangle==20230923
trimesh==4.4.1
util==0.0.1
vedo==2024.5.2
wandb==0.21.0
